requirements.txt
Django==4.2.7
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3

# REST API
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Image handling
Pillow==10.1.0

# Excel/PDF export
openpyxl==3.1.2
reportlab==4.0.7
xlsxwriter==3.1.9

# Email
django-anymail==10.2

# Security & Performance
django-ratelimit==4.1.0
django-environ==0.11.2

# Celery for async tasks (optional)
celery==5.3.4
redis==5.0.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3

# QR Code generation
qrcode==7.4.2
qrcode[pil]==7.4.2

# Charts and visualization
matplotlib==3.8.2
pandas==2.1.4

# Two-Factor Authentication
django-otp==1.3.0
qrcode==7.4.2

# Payment processing (Stripe example)
stripe==7.8.0

# Advanced caching
django-redis==5.4.0

# API documentation
drf-yasg==1.21.7

# Task scheduling
django-celery-beat==2.5.0

# Monitoring and logging
sentry-sdk==1.39.1

# Data validation
jsonschema==4.20.0
